Remove double-take-devel directory and update version
This commit removes the entire double-take-devel directory, including all its files such as Dockerfile, README.md, build.yaml, config.yaml, and icon.png. This directory was likely used for development purposes and is no longer needed in the main branch.

In addition, the version of the Double Take application in the double-take/config.yaml file has been updated from 1.13.11.5 to beta. This change might be due to the application moving from a stable release to a beta testing phase.