Update version and fix top menu bug
This commit updates the version of the Double Take application from 1.13.11.6 to 1.13.11.6.1 in various files including the Dockerfile and config.yaml. Additionally, a bug fix has been implemented for the top menu as indicated in the CHANGELOG.md. The Dockerfile has also been updated to reflect the new version of the application.